+++ src/app/app.component.ts
@@ -3,7 +3,7 @@ import { User } from './modules/login/models/user';
 import { Observable } from 'rxjs';
 import * as fromLogin from './modules/login/store/login.reducer';
 import { Store } from '@ngrx/store';
-import { LoginState } from './modules/login/store/login.reducer';
+import { AppState } from './store/app.reducer';
 
 @Component({
   selector: 'app-root',
@@ -15,7 +15,7 @@ export class AppComponent {
   public loggedUser$: Observable<User | null>;
 
   constructor(
-    public store: Store<LoginState>
+    public store: Store<AppState>
   ) {
     this.loggedUser$ = this.store.select(fromLogin.selectUser);
   }